+++ src/chatstyle/core/__init__.py
@@ -10,7 +10,9 @@ from .constants import (
 )
 from .errors import abort_if_force_without_tty, abort_if_missing_without_tty
 from .interactive import (
+    AUTO_PROMPT_ENV_VAR,
     InteractiveResolution,
+    auto_prompt_enabled,
     is_interactive_available,
     normalize_interactive,
     resolve_interactive_mode,
@@ -18,6 +20,7 @@ from .interactive import (
 
 __all__ = [
     "BACK_VALUE",
+    "AUTO_PROMPT_ENV_VAR",
     "CHECKBOX_SELECTED_INDICATOR",
     "CHECKBOX_UNSELECTED_INDICATOR",
     "FORCE_INTERACTIVE_NO_TTY_MESSAGE",
@@ -26,6 +29,7 @@ __all__ = [
     "MISSING_REQUIRED_NO_TTY_MESSAGE",
     "abort_if_force_without_tty",
     "abort_if_missing_without_tty",
+    "auto_prompt_enabled",
     "is_interactive_available",
     "normalize_interactive",
     "resolve_interactive_mode",